@@ -18,8 +18,8 @@
         		<artifactId>maven-compiler-plugin</artifactId>
         		<version>3.6.1</version>
         		<configuration>
-        			<source>1.8</source>
-        			<target>1.8</target>
+        			<source>17</source>
+        			<target>17</target>
         		</configuration>
 			</plugin>
 		</plugins>
@@ -32,7 +32,23 @@
 			<version>4.12</version>
 			<scope>test</scope>
 		</dependency>
-		
+		<dependency>
+			<groupId>org.projectlombok</groupId>
+			<artifactId>lombok</artifactId>
+			<version>1.18.24</version>
+			<scope>compile</scope>
+		</dependency>
+		<dependency>
+			<groupId>com.fasterxml.jackson.core</groupId>
+			<artifactId>jackson-databind</artifactId>
+			<version>2.17.0</version>
+		</dependency>
+		<dependency>
+			<groupId>org.junit.jupiter</groupId>
+			<artifactId>junit-jupiter-api</artifactId>
+			<version>5.10.2</version>
+			<scope>test</scope>
+		</dependency>
 		<dependency>
 		    <groupId>org.apache.commons</groupId>
 		    <artifactId>commons-collections4</artifactId>
